Changed word color to enum + bug fixes and new features

+++ lib/pages/words_bank_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:noam_learns_english/models/word.dart';
 import 'package:noam_learns_english/providers/words_provider.dart';
 import 'package:noam_learns_english/widgets/rating_buttons.dart';
@@ -43,14 +44,36 @@ class _WordsBankPageState extends State<WordsBankPage> {
                   trailing: IconButton(
                     icon: const Icon(Icons.delete),
                     onPressed: () {
-                      wordsProvider.deleteWord(index);
+                      showDialog(
+                        context: context,
+                        builder: (context) => AlertDialog(
+                          title: const Text('Delete Word'),
+                          content: const Text(
+                              'Are you sure you want to delete this word?'),
+                          actions: [
+                            TextButton(
+                              onPressed: () {
+                                Navigator.of(context).pop();
+                              },
+                              child: const Text('Cancel'),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                wordsProvider.deleteWord(index);
+                                Navigator.of(context).pop();
+                              },
+                              child: const Text('Delete'),
+                            ),
+                          ],
+                        ),
+                      );
                     },
                   ),
-                  tileColor: word.color == 'green'
+                  tileColor: word.color == WordColor.green
                       ? Colors.green.shade100
-                      : word.color == 'yellow'
+                      : word.color == WordColor.yellow
                           ? Colors.yellow.shade100
-                          : word.color == 'red'
+                          : word.color == WordColor.red
                               ? Colors.red.shade100
                               : Colors.white,
                   onTap: () {
@@ -59,55 +82,69 @@ class _WordsBankPageState extends State<WordsBankPage> {
                     showDialog(
                       context: context,
                       builder: (context) => AlertDialog(
-                        content: Form(
-                          key: _formKey1,
-                          child: Column(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
-                              TextFormField(
-                                validator: (value) =>
-                                    value!.isEmpty ? 'Required' : null,
-                                initialValue: _english,
-                                decoration: const InputDecoration(
-                                  labelText: 'English',
-                                ),
-                                onChanged: (value) {
-                                  _english = value;
-                                },
-                                textInputAction: TextInputAction.next,
+                        content: Column(
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            Form(
+                              key: _formKey1,
+                              child: Column(
+                                mainAxisSize: MainAxisSize.min,
+                                children: [
+                                  ElevatedButton(
+                                    style: ElevatedButton.styleFrom(
+                                      shape: const CircleBorder(),
+                                      padding: const EdgeInsets.all(10),
+                                    ),
+                                    onPressed: () {
+                                      wordsProvider.currentWord = word;
+                                      Navigator.of(context).pop();
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: const Icon(Icons.arrow_back),
+                                  ),
+                                  TextFormField(
+                                    validator: (value) =>
+                                        value!.isEmpty ? 'Required' : null,
+                                    initialValue: _english,
+                                    decoration: const InputDecoration(
+                                      labelText: 'English',
+                                    ),
+                                    onChanged: (value) {
+                                      _english = value;
+                                    },
+                                    textInputAction: TextInputAction.next,
+                                  ),
+                                  TextFormField(
+                                    validator: (value) =>
+                                        value!.isEmpty ? 'Required' : null,
+                                    textDirection: TextDirection.rtl,
+                                    initialValue: _hebrew,
+                                    decoration: const InputDecoration(
+                                        labelText: 'Hebrew'),
+                                    onChanged: (value) {
+                                      _hebrew = value;
+                                    },
+                                  ),
+                                  const SizedBox(height: 10),
+                                  ElevatedButton(
+                                    onPressed: () {
+                                      if (_formKey1.currentState!.validate()) {
+                                        wordsProvider.updateWord(
+                                            word.english,
+                                            Word(
+                                              english: _english,
+                                              hebrew: _hebrew,
+                                              color: word.color,
+                                            ));
+                                        Navigator.of(context).pop();
+                                      }
+                                    },
+                                    child: const Text('Save'),
+                                  ),
+                                ],
                               ),
-                              TextFormField(
-                                validator: (value) =>
-                                    value!.isEmpty ? 'Required' : null,
-                                textDirection: TextDirection.rtl,
-                                initialValue: _hebrew,
-                                decoration:
-                                    const InputDecoration(labelText: 'Hebrew'),
-                                onChanged: (value) {
-                                  _hebrew = value;
-                                },
-                              ),
-                              const SizedBox(height: 10),
-                              RatingButtons(
-                                  wordsProvider: wordsProvider,
-                                  currentIndex: index,
-                                  currentWord: word),
-                              ElevatedButton(
-                                onPressed: () {
-                                  if (_formKey1.currentState!.validate()) {
-                                    wordsProvider.updateWord(
-                                        index,
-                                        Word(
-                                          english: _english,
-                                          hebrew: _hebrew,
-                                        ));
-                                    Navigator.of(context).pop();
-                                  }
-                                },
-                                child: const Text('Save'),
-                              ),
-                            ],
-                          ),
+                            ),
+                          ],
                         ),
                       ),
                     );
@@ -156,13 +193,21 @@ class _WordsBankPageState extends State<WordsBankPage> {
                     ElevatedButton(
                       onPressed: () {
                         if (_formKey2.currentState!.validate()) {
+                          if (words.any((word) => word.english == _english)) {
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              const SnackBar(
+                                content: Text('Word already exists'),
+                              ),
+                            );
+                            return;
+                          }
                           wordsProvider.addWord(
                               Word(english: _english, hebrew: _hebrew));
                           _formKey2.currentState!.reset();
-                          FocusScope.of(context).nextFocus();
+                          FocusScope.of(context).previousFocus();
                         }
                       },
-                      child: const Text('Add Word'),
+                      child: const Text('Add'),
                     ),
                   ],
                 ),
